pyproject.toml
[project]
name = "agent-mcp"
version = "0.1.0"
description = "Space Domain Agent Platform using OpenAI Agents SDK and MCP servers for NASA, STAC, and Orbital Mechanics"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents",
    "python-dotenv",
    "asyncio",
    "rich",
    "httpx",
]
